move frm_main
